BETTERORIOKS-12: added PointsDisplay logic
