+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -116,6 +116,15 @@
         <item quantity="few">Окно %1$d минуты</item>
         <item quantity="other">Окно %1$d минут</item>
     </plurals>
+
+    <plurals name="weeks_before_event">
+        <item quantity="one">через %1$d неделю</item>
+        <item quantity="few">через %1$d недели</item>
+        <item quantity="other">через %1$d недель</item>
+    </plurals>
+    <string name="weeks_before_event_now">на этой неделе</string>
+    <string name="weeks_before_event_passed">прошло</string>
+
     <string name="week_number">%1$d - неделя</string>
     <string name="schedule_scroll_to_today">выбрать сегодняшний день в расписании</string>
     <string name="drop_down_menu">открыть выпадающее меню</string>
