+++ composeApp/src/commonMain/kotlin/model/subjectPerformance/DisplaySubjectPerformance.kt
@@ -1,0 +1,39 @@
+package model.subjectPerformance
+
+import androidx.compose.runtime.Composable
+import betterorioks.composeapp.generated.resources.Res
+import betterorioks.composeapp.generated.resources.weeks_before_event
+import betterorioks.composeapp.generated.resources.weeks_before_event_now
+import betterorioks.composeapp.generated.resources.weeks_before_event_passed
+import model.subjects.PointsDisplay
+import model.subjects.subjectsJson.jsonElements.Resource
+import org.jetbrains.compose.resources.pluralStringResource
+import org.jetbrains.compose.resources.stringResource
+
+sealed interface SubjectPerformanceListItem {
+
+    data class DisplayControlEvent(
+        val id: String,
+        val name: String,
+        val shortName: String,
+        override val currentPoints: String,
+        override val maxPoints: String,
+        val resources: List<Resource>,
+        val isBonus: Boolean
+    ) : PointsDisplay, SubjectPerformanceListItem
+
+    data class DisplayWeeksLeft(
+        val weeksLeft: Int
+    ) : SubjectPerformanceListItem {
+
+        @Composable
+        fun getWeeksLeftString(): String {
+            return when (weeksLeft) {
+                in Int.MIN_VALUE..<0 -> stringResource(Res.string.weeks_before_event_passed)
+                0 -> stringResource(Res.string.weeks_before_event_now)
+                else -> pluralStringResource(Res.plurals.weeks_before_event, weeksLeft, weeksLeft)
+            }
+        }
+
+    }
+}
