+++ composeApp/src/commonMain/kotlin/ui/subjectPerformanceScreen/SubjectPerformanceScreen.kt
@@ -1,4 +1,4 @@
-package ui.subjectPerformance
+package ui.subjectPerformanceScreen
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
